backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
openai-whisper==20240930
groq==0.4.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
httpx==0.27.2
aiofiles==24.1.0
